pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-company-crewai"
version = "0.1.0"
description = "V1 CrewAI multi-team organization scaffold for ai_company"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "crewai>=0.100.0",
  "PyYAML>=6.0.1",
  "pydantic>=2.6.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
